Ok, the fact that life sucks manifests itself in various ways.
Ok, the fact that life sucks manifests itself in various ways.

First of all GDK_INVERT and GDK_XOR gc functions do not do
what you want them to do, but what they are supposed to do, so
my cursor instead of being a nice black and firm box, is just
a pinkie thingie on my Indy.

Second, I got all of the grid draw logic wrong.  That is not
the way Excel does it, so tomorrow, I rewrite it.

Miguel.