Weee! Regular strings are now treated as values (so they can be used in
Weee!  Regular strings are now treated as values (so they
can be used in cell expressions).  This means you can actually
start typing expressions and they will actually make sense
(note that no recalc is yet supported).

The code to do interactive selections is there, but for some
reason the cursor is not paying much attention to me.

This feature lets you select a cell for an expression if you
type in a formula in the spreadsheet (just type '=' and use
the arrow keys to see my buggy code).