It autocalcs!
It autocalcs!

My technique is incredibly cool!  The engine should support
most excel features for recursive computations (we just need
the dialog boxes now).

It still has a wierdo crash from a mismatched string ref/unref
but that is going to be trivial to fix.

Next step:  Functions in Perl.