Redid the format parsing engine to be closer to the Excel formatting
Redid the format parsing engine to be closer to the Excel formatting
engine.  This basically is just a better front end to Clahey's number
rendering engine.

This now supports quoted strings in a format;  quoted characters;
spacing control formats;  color specification;  plus some ammount
of pre-parsing.

Getting it Excel compliant will be a bit more difficult than we
expected, as for example this format shows:

	000-0000-0000

should render 12345670987 as: 123-4567-0987