Two things that have been on my hard disk for a long time:
Two things that have been on my hard disk for a long time:

1. Comments-on-cells code is working.  We just need the nice dialog to
actually type something in.

It needs xml-io.c support though, hopefully Daniel will add this soon.

2. Dropped cell->entered_text.  We now only store the parsed version
of the cell (either as an ExprTree or as a Value).  This should help
mmeeks with his Excel-formula importing code, it should be possible
to do the right thing now instead of constructing a string representation
of the formula and re-parsing it.

Various fixes here and there as well.

Miguel

1998-11-10  Miguel de Icaza  <miguel@nuclecu.unam.mx>

	* src/sheet-view.c (sheet_view_comment_get_points): Scale properly
	the points to be in window cordinates.

	* src/sheet.c (sheet_col_set_width): Move comments on column
	change.

	* src/cell.c (cell_copy): Duplicated cells get col, row and sheet
	values set to NULL.

1998-11-05  Miguel de Icaza  <miguel@nuclecu.unam.mx>

	* src/fn-misc.c: Add ctype.h

	* src/sheet-view.c (sheet_view_comment_get_points): New utility
	routine used.
	(sheet_view_comment_relocate): New function used to relocate a
	comment canvas item in a cell to its new location.

	* src/cell.c (cell_relocate): Move the comments with the cell.
	(cell_realize, cell_unrealize, cell_comment_realize,
	cell_comment_unrealize): New functions used to create and destroy
	view-specific cell components (the comments are canvas objects
	that need to be created on a per-view basis).