Now the function draws the cell using the cell background color setting.
* src/cell.c (cell_draw): Now the function draws the cell
using the cell background color setting. Although it does
the job, I guess there must be a better way to do it. Thanks
to Michael for the hint.

* src/item-grid.c (item_grid_draw_cell): Commented out the lines
involved with patterns display in the cell, for that code is
not ready yet.

* src/dialog-cell-format (create_coloring_page): Reactivated
the background selection panel of the coloring page.

* src/dialog-cell-format (create_background_radio): Commented out
the addition of the pattern selection for a cell.