Make the graphics code link, and works. This does not mean it is trivial
Make the graphics code link, and works.
This does not mean it is trivial to get it to link.
Do not fix.  I will be looking at the problem "soon".

Well, it does not quite work.  The code is not yet linking the
selection correctly to the range vectors.

Miguel.