Bug 30247 This was subtle.
Bug 30247
This was subtle.

The problem was that we were not recognizing the input string '12345678901'
as a General number.  So it was being matched by the regexp for '0' and
formatting as such.  While that does point out a weakness in our number
matching
    : why are we bothering to compare against '0' if it is the same as
      'General').
The real question was
    : Why was strtod failing ?
The answer was that strol was setting errno, but strtod was not clearing it.
So we are now more careful about clearing it.

2000-10-31  Jody Goldberg <jgoldberg@home.com>

	* src/value.c (value_new_from_string) : Reset errno on over/underflow.
	* src/number-match.c (format_match_simple) : Ditto.

	* src/parser.y (yylex) : reset errno in the event of over/underflow.
	  test for 'e' as an indicator of flaoting point numbers.