Use glib's seek enum instead of the enum which used to be in libgsf.
* src/stf.c (stf_read_default_probe): Use glib's seek enum instead
of the enum which used to be in libgsf.

* plugins/applix/boot.c (applix_file_probe): Use glib's seek enum
instead of the enum which used to be in libgsf.

* plugins/excel/ms-biff.c (ms_biff_query_peek_next, ms_biff_put_var_seekto,
ms_biff_put_var_commit, ms_biff_put_len_commit): Use glib's seek
enum instead of the enum which used to be in libgsf.

* plugins/excel/ms-excel-write.c (biff_boundsheet_write_last,
  write_index): Ditto.

* plugins/lotus/boot.c (lotus_file_probe): Use glib's seek enum
instead of the enum which used to be in libgsf.

* plugins/lotus/lotus.c (record_peek_next): Ditto.

* plugins/plan-perfect/pln.c (pln_file_probe): Use glib's seek enum
instead of the enum which used to be in libgsf.

* plugins/qpro/qpro-read.c (qpro_check_signature, qpro_read_workbook):
Use glib's seek enum instead of the enum which used to be in libgsf.

* plugins/sc/sc.c (sc_file_probe): Use glib's seek enum instead of the
enum which used to be in libgsf.

* plugins/sylk/sylk.c (sylk_file_probe): Use glib's seek enum instead
of the enum which used to be in libgsf.

* plugins/xbase/xbase.c (record_seek, xbase_field_new): Use glib's seek enum
instead of the enum which used to be in libgsf.