Function: add hidden option to turn on localized function names.
This isn't ready for prime time yet.  When turned on, function names
ought to be translated in the gui.  That, of course, requires that
translators have translated the function names.  The place to do that
is in the help text.  For example:

msgid "PRODUCT:product  of the given values"
msgstr "PRODUKT:produktet af de givne værdier"

(Does that really have two spaces?)