xlsx: fix manual position of chart titles.
This fixes manual position of chart titles.  That isn't, by itself,
important, but we used to interpret the manual position as being for
the chart itself.

The whole text-in-chart handling could use a re-think.  It is currently
too much of a hack.  I believe the cause is a desire to poke the text
and attributes directly into an object.