xlsx: parser fixes.
Many attributes, notably enum attributes, have default values.  The defaults
may or may not agree with our defaults, so make things explicit.

This also gets away from assuming that the first attribute is "val",
even if the schema says that is the only possibility.  Xml allows
xmlns:foo attributes down the whole tree, so we cannot simply assume
that "val" comes first.

Handle the posibility of a themed colour with value (GOColor)0.