xlsx: make shared strings parsing more robust.
The spec isn't clear on the precise format of references to the shared strings
table.  Until now we have only seen integers, but #765544 shows a sample with
extra spaces.

Extra spaces are silly.  The point of the shared strings table is to reduce
file size and extra spaces work against that.  Nevertheless, it shouldn't
hurt to support that.

The file probably was not created by Excel.