expr-deriv: new code for analytic derivative of spreadsheet expressions.
This can generate some fairly impressive expressions for the derivative
of the function that an expression represents.

It is not particularly smart: it understands constants, +, -, *, /, ^,
and the functions sum and sumsq.  For anything else, it bails.  In case
you are wondering, those elements are enough to compute the square
residuals of rational models.