xlxs: improve export of comments.
This adds a legacy drawing for each comment.  ("legacy"?  "You keep using
that words.  I do not think it means what you think it means.")

This also adds xmlns:space="preserve" as needed for the cell comment
contents.  Both Excel and LO do that, but this causes validation against
the schema to fail.  I am still investigating what to do about that.