Arrays: code cleanup.
This moves us closer to making array corners and elements part of GnmExprTop
as opposed to the current state where they are in GnmExpr, but are only
allowed at top level.

Specifically, with this commit, we now query attributes of array
corners and elements mostly with a texpr.  Mostly.

A nice side effect is that many expr-impl.h includes can be removed.  We
have too many of those.