Editing: avoid excess precision for percentages.
This isn't a complete fix, but close enough.

The question we should answer is "what string representing 100*f is
the shortest, such that when converted to a double and divided by 100
results in f?"

Close enough here is that we should at least get this right for
reasonable sized numbers (<10000%) with a most 6 decimals.