Tests: Ensure we load the in-tree introspection overrides.
The old method of just setting the python load path does not work anymore.

We now hook into low levels of python to make sure that any search for
the overrides finds the in-tree copy.

This isn't elegant, but introspection doesn't seem to be built with
testing in mind.  Any python doesn't make it easy to substitute one
file for another.