css: implement #ifdef for built-in css.
This is a poor man's preprocessor.  A very poor man's.  It understands
basically two things at the moment:

|   #ifdef DARK
|     ...
|   #else
|     ...
|   #endif

and

|   #if GTK_CHECK_VERSION(...)
|     ...
|   #else
|     ...
|   #endif

Note, that "DARK" is set iff running under a dark theme.  The gtk version
check is a run-time check.

Note: this code is not in play when ~/.config/gtk-3.0/gtk.css is being read.