Columns/Rows: performance improvements.
See bug #692.

Performance problems were observed for copying a large number of
cell comments.  Part of that problem was measuring the pixel
width/height of large ranges of columns/rows.

The old design was linear leading to O(n^2) behaviour when used in
a loop.

This design caches the pixel position of each segment [block of
128 columns/rows].  With that, the amortized cost is constant.

It isn't clear that the current ColRowCollection design is right,
but this patch will lead a bit of life to the current design.