Style: redo common style interface in terms of GPtrArray.
Using a container that knows its length makes this introspectable.
At the same time also hold a ref for every style.  Most callers
do not actually need that -- for file format writers there is no
chance that the style disappears underneath the caller -- but it
doesn't hurt in that situation either.