Compilation: warning fix.
We generally do not allocate enough space for a CellTile, just for the
relevant subtype.  That's mildly naughty when we turn, say, a
CellTileSimple* to a CellTile*.

With this patch we use a flexible array member for the pointers of
the "any" member instead of just using a CellTileMatrix.  The
flexible member is actually the right thing to do so the compiler
stays out of the indexing check business.