Fix zoom problem.
This works around a change in Pango with respect to setting attributes
on layouts and what is supposed to happen when the attribute list being
set is later changed.  I hesitate calling it an API break -- it
technically is -- because we should probably not have been relying on
the behaviour.